Beta 5 and SwityTimer
made a few changes to run on beta 5
  - changed picker.focusForCrownInput() to picker.focus()
  - delete the BallControllerDelegate’s @objc tag and all optional tags
(wouldn’t build without it and its nicer that way)

in also copied the JonyHead and Borrito image for @1x and @2x because
they weren’t visible on my setup.

finally I added the SwiftyTimer extension and changed all timers to use
closures.